move greet players and prompt players to display class
