+++ lib/display_double.rb
@@ -1,6 +1,19 @@
+require "game"
+
 class DisplayDouble
   def initialize(moves)
     @moves = moves
+  end
+
+  def greet_players
+    puts "Welcome to Tic Tac Toe"
+    puts "To make a move enter the number of the position you want to play"
+    puts "Player1's mark is 'X' and Player2's mark is 'O'"
+    puts "The game will end either when a player wins by placing their mark across a row, column or diagonally, or all the positions are taken\n"
+  end
+
+  def prompt_player
+    print "\nPlayer #{game.current_player}, make a move: "
   end
 
   def get_move(board)
